Fix lint config and organize models
